clarify naming, analysis tests
